자소서 메타정보+질문지+본문을 챗지피티에게 보내는 프로세스 구현중

+++ GIT_PROJECT/src/main/java/com/itwillbs/project/review/dto/CoverLetterDTO.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @ToString
 public class CoverLetterDTO {
 	
-	private Integer coverLetterIdx;	// 자소서 일련번호
+	private Long coverLetterIdx;	// 자소서 일련번호
 	private Long userId; 			// 사용자 식별자(user(user_id))
 	
 	private String title;			// 자소서 제목
